Move room related events to its own file
