+++ tsconfig.json
@@ -7,6 +7,7 @@
     "moduleResolution": "node",
     "sourceMap": true,
     "strict": true,
+    "strictPropertyInitialization": false,
     "target": "esnext",
     "paths": {
       "*": ["node_modules/*", "src/types/*"]
